s3:vfs: change files_struct.fnum from int to uint64_t

Autobuild-User(master): Stefan Metzmacher <metze@samba.org>
Autobuild-Date(master): Fri Jun 15 05:22:41 CEST 2012 on sn-devel-104

+++ source3/include/vfs.h
@@ -185,7 +185,7 @@ struct fd_handle {
 
 typedef struct files_struct {
 	struct files_struct *next, *prev;
-	int fnum;
+	uint64_t fnum;
 	struct connection_struct *conn;
 	struct fd_handle *fh;
 	unsigned int num_smb_operations;
